% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_datasets.R
\name{clinical.20160128}
\alias{ACC.clinical.20160128}
\alias{ACC.clinical.20160128,BLCA.clinical.20160128,BRCA.clinical.20160128,CESC.clinical.20160128,CHOL.clinical.20160128,COADREAD.clinical.20160128,DLBC.clinical.20160128,ESCA.clinical.20160128,FPPP.clinical.20160128,GBMLGG.clinical.20160128,HNSC.clinical.20160128,KICH.clinical.20160128,KIPAN.clinical.20160128,KIRC.clinical.20160128,KIRP.clinical.20160128,LAML.clinical.20160128,LIHC.clinical.20160128,LUAD.clinical.20160128,LUSC.clinical.20160128,MESO.clinical.20160128,OV.clinical.20160128,PAAD.clinical.20160128,PCPG.clinical.20160128,PRAD.clinical.20160128,SARC.clinical.20160128,SKCM.clinical.20160128,STAD.clinical.20160128,STES.clinical.20160128,TGCT.clinical.20160128,THCA.clinical.20160128,THYM.clinical.20160128,UCEC.clinical.20160128,UCS.clinical.20160128,UVM.clinical.20160128}
\alias{BLCA.clinical.20160128}
\alias{BRCA.clinical.20160128}
\alias{CESC.clinical.20160128}
\alias{CHOL.clinical.20160128}
\alias{COADREAD.clinical.20160128}
\alias{DLBC.clinical.20160128}
\alias{ESCA.clinical.20160128}
\alias{FPPP.clinical.20160128}
\alias{GBMLGG.clinical.20160128}
\alias{HNSC.clinical.20160128}
\alias{KICH.clinical.20160128}
\alias{KIPAN.clinical.20160128}
\alias{KIRC.clinical.20160128}
\alias{KIRP.clinical.20160128}
\alias{LAML.clinical.20160128}
\alias{LIHC.clinical.20160128}
\alias{LUAD.clinical.20160128}
\alias{LUSC.clinical.20160128}
\alias{MESO.clinical.20160128}
\alias{OV.clinical.20160128}
\alias{PAAD.clinical.20160128}
\alias{PCPG.clinical.20160128}
\alias{PRAD.clinical.20160128}
\alias{SARC.clinical.20160128}
\alias{SKCM.clinical.20160128}
\alias{STAD.clinical.20160128}
\alias{STES.clinical.20160128}
\alias{TGCT.clinical.20160128}
\alias{THCA.clinical.20160128}
\alias{THYM.clinical.20160128}
\alias{UCEC.clinical.20160128}
\alias{UCS.clinical.20160128}
\alias{UVM.clinical.20160128}
\alias{clinical.20160128}
\title{clinical datasets from TCGA project from 2016-01-28 release date}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.clinical.20160128(metadata = FALSE)

BLCA.clinical.20160128(metadata = FALSE)

BRCA.clinical.20160128(metadata = FALSE)

CESC.clinical.20160128(metadata = FALSE)

CHOL.clinical.20160128(metadata = FALSE)

COADREAD.clinical.20160128(metadata = FALSE)

DLBC.clinical.20160128(metadata = FALSE)

ESCA.clinical.20160128(metadata = FALSE)

FPPP.clinical.20160128(metadata = FALSE)

GBMLGG.clinical.20160128(metadata = FALSE)

HNSC.clinical.20160128(metadata = FALSE)

KICH.clinical.20160128(metadata = FALSE)

KIPAN.clinical.20160128(metadata = FALSE)

KIRC.clinical.20160128(metadata = FALSE)

KIRP.clinical.20160128(metadata = FALSE)

LAML.clinical.20160128(metadata = FALSE)

LIHC.clinical.20160128(metadata = FALSE)

LUAD.clinical.20160128(metadata = FALSE)

LUSC.clinical.20160128(metadata = FALSE)

MESO.clinical.20160128(metadata = FALSE)

OV.clinical.20160128(metadata = FALSE)

PAAD.clinical.20160128(metadata = FALSE)

PCPG.clinical.20160128(metadata = FALSE)

PRAD.clinical.20160128(metadata = FALSE)

SARC.clinical.20160128(metadata = FALSE)

SKCM.clinical.20160128(metadata = FALSE)

STAD.clinical.20160128(metadata = FALSE)

STES.clinical.20160128(metadata = FALSE)

TGCT.clinical.20160128(metadata = FALSE)

THCA.clinical.20160128(metadata = FALSE)

THYM.clinical.20160128(metadata = FALSE)

UCEC.clinical.20160128(metadata = FALSE)

UCS.clinical.20160128(metadata = FALSE)

UVM.clinical.20160128(metadata = FALSE)
}
\arguments{
\item{metadata}{A logical indicating whether load data into the workspace (default, \code{FALSE}) or to only display the object's metadata (\code{TRUE}). See examples.}
}
\description{
Package provides clinical datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the release date: \code{2016-01-28} . Visit \pkg{RTCGA} site: \url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.clinical.20160128}  can be found here: \code{browseVignettes("RTCGA")}. Link to the data format explanation is in the package DESCRIPTION file.
}
\details{
\code{browseVignettes("RTCGA")}
}
\examples{

\dontrun{
 ACC.clinical.20160128(metadata = TRUE) 
 ACC.clinical.20160128(metadata = FALSE) 
 ACC.clinical.20160128 
}

}

